refactor(frontend): extract CampaignTable
